quartz: Align filtering quality with image backend
The Quartz framework provides several quality settings, but they were
not used. Now the translation of cairo_filter_t to Quartz filtering
modes tries to match the quality settings from image as much as
possible.

Specifically, CAIRO_FILTER_GOOD and CAIRO_FILTER_BILINEAR are now
converted to kCGInterpolationLow, which seems to be a bilinear filter.

Fixes:
 - pixman-downscale-bilinear-24
 - pixman-downscale-bilinear-95
 - pixman-downscale-good-24
 - pixman-downscale-good-95
 - pthread-same-source
 - recording-surface-extend-none
 - recording-surface-extend-reflect
 - recording-surface-extend-repeat
 - recording-surface-over
 - recording-surface-source
 - surface-pattern-big-scale-down
 - surface-pattern-scale-down
 - surface-pattern-scale-down-extend-none
 - surface-pattern-scale-down-extend-reflect
 - surface-pattern-scale-down-extend-repeat